requirements.txt
torch==2.0.1
pandas
scikit-learn
fastapi
pydantic
httpx
typer
omegaconf
hydra-core
wandb
streamlit
uvicorn
tqdm
prometheus-fastapi-instrumentator
prometheus-client
